Abstract QPSolvers as a 'pseudoclass' to handle choice of solver engine
Previously, QPSolver, MIQPSolver, QPSolver2, and MIQPSolver2 wrapped either CPLEX or OSQP packages, and presented the same interface.

First, we specify that required interface in a dummy _Base_QPSolver class, to clarify behaviour for future developers.
Then, we rename the QPSolver, MIQPSolver, QPSolver2, and MIQPSolver2 classes more descriptively.
Finally, we create "pseudoclasses" QPSolver and MIQPSolver that (a) return an appropriate, accessible solver class, (b) warn if the requested solver is not accessible, or (c) raise an ImportError if no solvers are available.