(6/) Refactor _run_qfit_instance to run on a single residue
Previously, _run_qfit_instance took a list of residues, and iterated over them independently.
This saved some overhead in starting up processes. However, it added code complexity for managing the queue of residues, for printing progress to the screen, and for holding a threadsafe Counter.

This patch slims _run_qfit_instance to a minimal task: find conformers of a single residue.