(10/) Finish off __init__ of residue objects passed to subprocess
This code is copied from the qfit.structure.base_structure._BaseStructure.__init__ constructor.
I am not sure why this has not appeared as an issue in the previous version of the code—
though I suspect there is a difference in the way these objects are passed to
mp.Process or the mp.Process spawned by the mp.Pool.