Make version.py not crash when it can't find tags
When a git repo doesn't contain tags, version.py crashes, and setup.py can't continue.
This is a problem if someone does a shallow clone on the repo (i.e. with --depth=1). This happens in the GitHub CI runner, causing issues for testing.

This patch fixes it by returning the commit SHA if it can't find a tag.