Use pip & pyproject.toml to install qFit, update README
numpy _has_ to be installed before `python setup.py install`.
This is because in `setup.py`, there is an import requirement for numpy; we use numpy to locate the numpy headers, so that we can compile our C extension.
This is still the case.

However, pip can manage the order in which these requirements are satisfied.
By adding pyproject.toml (PEP 518), we can direct pip to install numpy before building setup.py.
PEP 518: if pyproject.toml is present, pip will create a build environment, and use this.

With this patch, we make sure pip installs numpy into the build environment---this ensures it is present before we attempt to compile extensions.

This patch updates the README to specify that this is the preferred installation method.
`conda` is able to work with pip, and will manage the environment that pip installs into.

We also update the GitHub CI workflow, it should use pip to install too.

Pip (through pyproject.toml) can handle the build prerequisite of numpy.