Use setuptools_scm to manage package versioning (not version.py)
Things in the root directory (e.g. setup.py, version.py) are not bundled into a wheel with the qfit package.
On installation into a virtualenv (as is done during testing), version.py will not be present.
Attempting to import it thus presents a problem.

We replace the whole script with a well-managed package, setuptools_scm to sidestep this issue.

Side effect: versioning numbers change a little, but is now conformant with semantic versioning  (à la PEP 440).