Add tox to GitHub test runner
No tests should be run in the conda "base" environment in the GitHub runner.
As such, we can remove a lot of package installs in the runner.

Instead tox-conda will manage the creation of appropriate conda envs, and then perform testing within these.

We can make tox construct multiple conda envs (in the future), e.g. one for testing behaviour with CPLEX, one for testing behaviour with MIOSQP.