(9/) Force multiprocessing to not use default 'fork'
On macOS (Unix-based), I've come across a locking issue, after MemoryErrors occur.
I believe this to be a race issue on the forking of a new Pool process: when one of the processes is killed because of a MemoryError, a new process is forked from the master, but it is forked from a state where the lock has already been acquired (see https://codewithoutrules.com/2018/09/04/python-multiprocessing/ for a better explanation).

I imagine the previous method never encountered this error, since all processes were only spawned once.

Spawn (fork+execve)... should fix this issue, since the process state is reinitialised?
I don't think we are really worried about the slower overhead on process startup.