Use numpy methods, not nested loops
This commit speeds up all of qFit by a factor of ~2.3.

Previously, _QPSolver.initialize used nested for-loops to compute inner products.
Given that this came from a C++ project, I understand why it was there.
But, Python can't access the same loop-unrolling optimisations of a compiled language, so it's mighty slow... not to mention all the array slicing and access overhead.

Previously, this routine took around 140 s (just to construct an array!). It now takes about 4.6 s.

By default, np.inner will do a sum product over the last axes: so for the 2D matrices we have here, np.inner(A[i,j], B[k,l]) -> C[i,k] . This was precisely the function of the nested for loops.