Make CBAngleRotator act on Cα-Cβ-Cγ atoms
The intention of CBAngleRotator _should_ be to deflect aromatic rings
of Trp, Tyr, Phe, His residues. This is best done by flexing the
Cα-Cβ-Cγ angle.
We expect that for these aromatic residues, the computational trade-off
will be worth it: this extra sampling of a d.o.f. will improve density
matching. We do not expect this trade-off to be worth it for other
residue types.

This commit:
* Shifts the angle flexion from N-Cα-Cβ to Cα-Cβ-Cγ
* Limits QFitRotamericResidue._sample_angle to only run on
  {Trp, Tyr, Phe, His}
* Adds documentation for samplers.CBAngleRotator
  * Explains rationale for restriction of operation to certain residues