(7c/) Use dicts to represent pdbfile records (NB: py>=3.6)
The previous implementation used tuples, but the fields being accessed were unclear.

Namedtuples fixed the clarity problem, but they are immutable.
We often need to replace fields in the record, so this is not fantastic.

Instead, this commit uses dict to bring 'clarity' and allow 'mutability'.

---- NB: py>=3.6 ----
Through PEP468, a CPython 3.6 implementation detail effectively made
dictionary order == insertion order.

This was codified in the language standards in CPython 3.7.

We rely on this ordering being preserved when we unpack record.values().
---------------------