[REVERT] (3/) Use pytest and coverage from external to the tox env
No, this was a bad idea. (Except for Windows, which got a bit further!)

Current guess: the package is being built by the runner to a wheel.
This wheel has all the DLL dependencies from the base environment.
TOX then installs this AS A WHEEL, and doesn't have the right DLL links to the venv numpy.

P.S. Not sure if I believe this anymore? This commit is out of order.
When this was running earlier, we were getting "could not delete vs_blah" when the tox env was being created.
This commit got us to the "numpy couldn't import modules from dlls" errors.