Change all store_false actions with --no- to ToggleActionFlag type
This is the crux of this PR: there's too many negatives floating around here.

We fold the 'store_false' action and --no- flags together: they cancel each other out.
We get a positively worded flag, and make sure to set the default to True.
Finally, we assign the ToggleActionFlag type so that both options appear together in the help.

Note, we lose some short flags as a result (I don't think this is a loss for clarity).

What was
  -bb, --no-backbone        Do not sample backbone using inverse kinematics. (default: True)
is now:
  --[no-]backbone      Sample backbone using inverse kinematics. (default: True)

This applies to:
* -ns, --no-scale → --[no-]scale
* -nosub, --no-subtract → --[no-]subtract
* -bb, --no-backbone → --[no-]backbone
* -sa, --no-sample-angle → --[no-]sample-angle
* -T, --no-threshold-selection → --[no-]threshold-selection
* -Ts, --no-segment-threshold-selection → --[no-]segment-threshold-selection