Strip amide H before calculating RMSF (necessary b/c of backbone-H bug)
phenix.refine creates a consistent backbone between residues i-1 and i.
This means that some altconfs in residue i end up having different
numbers of atoms (due to a bonus amide H from the corresponding altconf
in residue n-1).
We strip "name H" atoms from the structure to sidestep this error (this
is unlikely to change the magnitude of the measured RMSF much).

Kinda hacky.