Strip all H before calculating RMSF
An extension to 48b69fa9.

Unfortunately, we can't guarantee that His will maintain a consistent
protonation state between altconfs (reduce assesses them separately,
phenix.refine respects this).

We strip "e. H" atoms from the structure to sidestep this error (this
is unlikely to change the magnitude of the measured RMSF much).