Use --no-build-isolation install in CI script (fix numpy C-API crash)
Without this option, pip will go fetch all the install_requires listed
in setup.py for an isolated build. The problem is that the version of
numpy it grabs could be different to one that is _already_ present in
the conda env.

Since numpy _is_ already present in the conda env, we expose it to pip
using the --no-build-isolation flag. This way, our C extension is built
against the same numpy as will be present in the conda env.

This is particularly important given numpy 1.20 has recently shifted
API version (see numpy/numpy/core/include/numpy/numpyconfig.h,
numpy/numpy/core/code_generators/cversions.txt).