Limit numpy upper version in both _build_ and _install_ sections
First, note that the requirements for the build environment (e.g. for extensions)
  is different to the requirements for running the package (i.e. install_requires).
Build requirements are listed in pyproject.toml, install requirements in setup.py.

We don't need --no-build-isolation if we specify the same numpy requirements
  in both.
As such, I have reverted the previous commit.

There is a consistent numpy C-API between numpy 1.17 (technically 1.16.1)
  and 1.20 (see previous commit).
This commit consistently limits the numpy versions to that range.