Add a function (and test!) for normalizing arrays to a fixed no. of d.p.
IMPLEMENTATION
First normalize to 1.0 as a float, round elements to n dp.
If the sum of the array isn't 1.00..0 yet, add the appropriate number
of rounding epsilons (e.g. 0.01) to the members of the array from largest
to smallest. Return the array in the order it was provided.

NOTES
- This includes the package 'hypothesis' (required for testing only!), so
  I've tried to add that where I think it belongs.
- I doubt this is very efficient. There's probably a smarter way?

TODO
- Add calls to this function whenever normalizing an array to 1.00.
  For now, this is when (re)normalizing conformer occupancies.