Enforce Docker x86_64 emulation
Running vscode on an M1 Mac, Docker runs linux_arm64 by default.
CPLEX (and mkl) are available on the conda Intel repos only.
Conda will fail to find CPLEX (and the mkl version of libblas), and the
  container fails to build.

Putting --platform=linux/amd64 in the Dockerfile forces the "Remote
Container" to use x86_64 emulation. We've already put a similar
workaround in the README.md for the main package (modifying conda
instructions for users on M1 Macs), so this should put devs in a
similar DevContainer as a default User environment.