adding reduce failure flag
As we check for rotamer for NQH residues in qFit, it is less likely we need to optimize their positions during refinement. Reduce takes an enormous amount of memory, causing refinement to crash. This change will allow the refinement to run without checking for these flips and tell the user if it has run refinement without checking for these flips.