Calculate BIC 'k' as per paper
Previously, there was a try/except block that calculated 'k' in different
ways. In none of them did we divide by 'tdmin'.

From the paper (10.1002/pro.4001): k=4*n(active atoms)/tdmin .
The factor 1/tdmin is a proxy for model complexity and imposes a limit
on the maximum number of conformations.