Select the number of conformers that minimizes the BIC
We say we do this in the paper (10.1002/pro.4001) --- it is a
reasonable approach to 'parsimony' in the qFit routine.

(The number of conformers is _drastically_ decreased after this.)

Yet, since the assignment to occupancies occurred _outside_ the for loop,
and 7e2b5e2 removed the 'break' from the for loop,
we ended up just selecting the number of conformers from the _last_ iteration
of the BIC loop. Always.

History of _solve() routine:
- the _solve() routine was first written as commented code in 136357f
- it was uncommented in 2a9f454
- the BIC loop break was removed in 7e2b5e2 (source of the problem)
- 'k' was introduced to the BIC function in f97a090