updating missing backbones
We have been looking at the entire segment (often the entire chain or protein) for missing backbone atoms. If there are any missing backbones, all backbone sampling for that residue is abandoned. We should only check if the backbone exists within the segment used for the inverse kinematic algorithm (residue - 3 or residue +3).