BIC: Update k formula, don't _estimate_ n(altconf atoms), we _know_ this
1/threshold is the maximum number of conformers that will be returned by the solver (i.e. an overestimate).
As such, natoms/threshold will always overestimate of the total number of atoms in the model.

Since we can calculate the total number of conformers returned after the solver (nconfs)
we can use this to calculate k, rather than estimating it.

This reverts k to how it was in f97a090.