update qfit.py
make sure _write_intermediate_conformers works for running ligand. I had to make a slight change to the _write_intermediate_conformers function to include an optional argument for inputting a coordinate array. This was necessary in order to write out the conformers for the different ligand sampling functions separately. The function can still be used without this argument, and will just write out confs stored in self._coor_set, as it already does